training/requirements.txt
tensorflow>=2.14
tensorflow-metal
tensorflowjs>=4.0
Pillow
numpy
